Initialize Card service in lifecycle protocol
I used dosync and refs here because assoc-ing the state onto the record
was simply returning a new record (because immutability) instead of
modifying it in place, and thus the defrpc wasn't getting the state it
needed for get-state*. So I'm (ab)using the STM to mutate this in
place.

Unfortunately, this doesn't quite get me the component-reusability that
I truly want. I'd like to be able to create multiple instances of the
component for different Trello lists, but that would mean each instance
of the component would attempt to define a get-state RPC, and the
definitions would clash.

I guess the way to solve this would be to make the component take a
sequnce of list-ids, and make needs-funding-cards multivariadic. I'd
also like to use a scheduler to frequently call out to the Trello API,
instead of just doing it on startup.